requirements.txt
tweepy
openai
python-dotenv
pytz